requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
click
rich
pykrx
yfinance
websockets
openai
pdfplumber>=0.11
feedparser
sqlalchemy
alembic
pytest
pytest-asyncio
apscheduler
httpx>=0.27.0
psycopg2-binary
pyyaml
bcrypt
python-jose[cryptography]
cryptography>=42